event handlers, region completion
esssentially encapsulated everything, finishing the process of having regions notify manager when complete.
need to handle out of bounds error, the complete the animation gueue from each region, should probably come up with an easy way to see which regions failed or not. and if a region fails, it shouldnt impede the next region, it just woulndt affect the updated tile possibilites