found new problem
I believe that the setnextcell function goes straight to the cell array of the grid, so maybe i just need to offset where it sets it on the tile map? getting gross
I think that cells only need to be offset before animating them on the tilemap in setnext cell. should rework region management ingeneral tho